Roll Shared Z Axis Transitions into the same builder as rest of SharedAxisTransition types
