+++ packages/animations/lib/animations.dart
@@ -6,4 +6,3 @@ export 'src/fade_through_transition.dart';
 export 'src/open_container.dart';
 export 'src/page_transition_switcher.dart';
 export 'src/shared_axis_transition.dart';
-export 'src/shared_z_axis_transition.dart';
