+++ packages/animations/test/shared_axis_transition_test.dart
@@ -970,6 +970,398 @@ void main() {
       },
     );
   });
+
+  group('SharedAxisTransitionType.scaled', () {
+    testWidgets(
+      'SharedAxisPageTransitionsBuilder builds a SharedAxisTransition',
+      (WidgetTester tester) async {
+        final AnimationController animation = AnimationController(
+          vsync: const TestVSync(),
+        );
+        final AnimationController secondaryAnimation = AnimationController(
+          vsync: const TestVSync(),
+        );
+
+        await tester.pumpWidget(
+          const SharedAxisPageTransitionsBuilder(
+            transitionType: SharedAxisTransitionType.scaled,
+          ).buildTransitions<void>(
+            null,
+            null,
+            animation,
+            secondaryAnimation,
+            const Placeholder(),
+          ),
+        );
+
+        expect(find.byType(SharedAxisTransition), findsOneWidget);
+      },
+    );
+
+    testWidgets(
+      'SharedAxisTransition runs forward',
+      (WidgetTester tester) async {
+        final GlobalKey<NavigatorState> navigator = GlobalKey<NavigatorState>();
+        const String bottomRoute = '/';
+        const String topRoute = '/a';
+
+        await tester.pumpWidget(
+          _TestWidget(
+            navigatorKey: navigator,
+            transitionType: SharedAxisTransitionType.scaled,
+          ),
+        );
+
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 1.0);
+        expect(_getOpacity(bottomRoute, tester), 1.0);
+        expect(find.text(topRoute), findsNothing);
+
+        navigator.currentState.pushNamed(topRoute);
+        await tester.pump();
+        await tester.pump();
+
+        // Bottom route is full size and fully visible.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 1.0);
+        expect(_getOpacity(bottomRoute, tester), 1.0);
+        // Top route is at 80% of full size and not visible yet.
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 0.8);
+        expect(_getOpacity(topRoute, tester), 0.0);
+
+        // Jump 3/10ths of the way through the transition, bottom route
+        // should be be completely faded out while the top route
+        // is also completely faded out.
+        // Transition time: 300ms, 3/10 * 300ms = 90ms
+        await tester.pump(const Duration(milliseconds: 90));
+
+        // Bottom route is now invisible
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+        // Top route is still invisible, but scaling up.
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getOpacity(topRoute, tester), 0.0);
+        double topScale = _getScale(topRoute, tester);
+        expect(topScale, greaterThan(0.8));
+        expect(topScale, lessThan(1.0));
+
+        // Jump to the middle of fading in
+        await tester.pump(const Duration(milliseconds: 90));
+        // Bottom route is still invisible
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+        // Top route is fading in
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getOpacity(topRoute, tester), greaterThan(0));
+        expect(_getOpacity(topRoute, tester), lessThan(1.0));
+        topScale = _getScale(topRoute, tester);
+        expect(topScale, greaterThan(0.8));
+        expect(topScale, lessThan(1.0));
+
+        // Jump to the end of the transition
+        await tester.pump(const Duration(milliseconds: 120));
+        // Bottom route is not visible.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 1.1);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+        // Top route fully scaled in and visible.
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 1.0);
+        expect(_getOpacity(topRoute, tester), 1.0);
+
+        await tester.pump(const Duration(milliseconds: 1));
+        expect(find.text(bottomRoute), findsNothing);
+        expect(find.text(topRoute), findsOneWidget);
+      },
+    );
+
+    testWidgets(
+      'SharedAxisTransition runs in reverse',
+      (WidgetTester tester) async {
+        final GlobalKey<NavigatorState> navigator = GlobalKey<NavigatorState>();
+        const String bottomRoute = '/';
+        const String topRoute = '/a';
+
+        await tester.pumpWidget(
+          _TestWidget(
+            navigatorKey: navigator,
+            transitionType: SharedAxisTransitionType.scaled,
+          ),
+        );
+
+        navigator.currentState.pushNamed(topRoute);
+        await tester.pumpAndSettle();
+
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 1.0);
+        expect(_getOpacity(topRoute, tester), 1.0);
+        expect(find.text(bottomRoute), findsNothing);
+
+        navigator.currentState.pop();
+        await tester.pump();
+
+        // Top route is full size and fully visible.
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 1.0);
+        expect(_getOpacity(topRoute, tester), 1.0);
+        // Bottom route is at 80% of full size and not visible yet.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 0.8);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+
+        // Jump 3/10ths of the way through the transition, bottom route
+        // should be be completely faded out while the top route
+        // is also completely faded out.
+        // Transition time: 300ms, 3/10 * 300ms = 90ms
+        await tester.pump(const Duration(milliseconds: 90));
+
+        // Bottom route is now invisible
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getOpacity(topRoute, tester), 0.0);
+        // Top route is still invisible, but scaling up.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(
+          _getOpacity(bottomRoute, tester),
+          moreOrLessEquals(0, epsilon: 0.005),
+        );
+        double bottomScale = _getScale(bottomRoute, tester);
+        expect(bottomScale, greaterThan(0.8));
+        expect(bottomScale, lessThan(1.0));
+
+        // Jump to the middle of fading in
+        await tester.pump(const Duration(milliseconds: 90));
+        // Top route is still invisible
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getOpacity(topRoute, tester), 0.0);
+        // Bottom route is fading in
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getOpacity(bottomRoute, tester), greaterThan(0));
+        expect(_getOpacity(bottomRoute, tester), lessThan(1.0));
+        bottomScale = _getScale(bottomRoute, tester);
+        expect(bottomScale, greaterThan(0.8));
+        expect(bottomScale, lessThan(1.0));
+
+        // Jump to the end of the transition
+        await tester.pump(const Duration(milliseconds: 120));
+        // Top route is not visible.
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 1.1);
+        expect(_getOpacity(topRoute, tester), 0.0);
+        // Bottom route fully scaled in and visible.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 1.0);
+        expect(_getOpacity(bottomRoute, tester), 1.0);
+
+        await tester.pump(const Duration(milliseconds: 1));
+        expect(find.text(topRoute), findsNothing);
+        expect(find.text(bottomRoute), findsOneWidget);
+      },
+    );
+
+    testWidgets(
+      'SharedAxisTransition does not jump when interrupted',
+      (WidgetTester tester) async {
+        final GlobalKey<NavigatorState> navigator = GlobalKey<NavigatorState>();
+        const String bottomRoute = '/';
+        const String topRoute = '/a';
+
+        await tester.pumpWidget(
+          _TestWidget(
+            navigatorKey: navigator,
+            transitionType: SharedAxisTransitionType.scaled,
+          ),
+        );
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(find.text(topRoute), findsNothing);
+
+        navigator.currentState.pushNamed(topRoute);
+        await tester.pump();
+
+        // Jump to halfway point of transition.
+        await tester.pump(const Duration(milliseconds: 150));
+        // Bottom route is fully faded out.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+        final double halfwayBottomScale = _getScale(bottomRoute, tester);
+        expect(halfwayBottomScale, greaterThan(1.0));
+        expect(halfwayBottomScale, lessThan(1.1));
+
+        // Top route is fading/scaling in.
+        expect(find.text(topRoute), findsOneWidget);
+        final double halfwayTopScale = _getScale(topRoute, tester);
+        final double halfwayTopOpacity = _getOpacity(topRoute, tester);
+        expect(halfwayTopScale, greaterThan(0.8));
+        expect(halfwayTopScale, lessThan(1.0));
+        expect(halfwayTopOpacity, greaterThan(0.0));
+        expect(halfwayTopOpacity, lessThan(1.0));
+
+        // Interrupt the transition with a pop.
+        navigator.currentState.pop();
+        await tester.pump();
+
+        // Nothing should change.
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), halfwayBottomScale);
+        expect(_getOpacity(bottomRoute, tester), 0.0);
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), halfwayTopScale);
+        expect(_getOpacity(topRoute, tester), halfwayTopOpacity);
+
+        // Jump to the 1/4 (75 ms) point of transition
+        await tester.pump(const Duration(milliseconds: 75));
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), greaterThan(1.0));
+        expect(_getScale(bottomRoute, tester), lessThan(1.1));
+        expect(_getScale(bottomRoute, tester), lessThan(halfwayBottomScale));
+        expect(_getOpacity(bottomRoute, tester), greaterThan(0.0));
+        expect(_getOpacity(bottomRoute, tester), lessThan(1.0));
+
+        // Jump to the end.
+        await tester.pump(const Duration(milliseconds: 75));
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(_getScale(bottomRoute, tester), 1.0);
+        expect(_getOpacity(bottomRoute, tester), 1.0);
+        expect(find.text(topRoute), findsOneWidget);
+        expect(_getScale(topRoute, tester), 0.80);
+        expect(_getOpacity(topRoute, tester), 0.0);
+
+        await tester.pump(const Duration(milliseconds: 1));
+        expect(find.text(topRoute), findsNothing);
+        expect(find.text(bottomRoute), findsOneWidget);
+      },
+    );
+
+    testWidgets(
+      'SharedAxisTransition properly disposes animation',
+      (WidgetTester tester) async {
+        final GlobalKey<NavigatorState> navigator = GlobalKey<NavigatorState>();
+        const String bottomRoute = '/';
+        const String topRoute = '/a';
+
+        await tester.pumpWidget(
+          _TestWidget(
+            navigatorKey: navigator,
+            transitionType: SharedAxisTransitionType.scaled,
+          ),
+        );
+        expect(find.text(bottomRoute), findsOneWidget);
+        expect(find.text(topRoute), findsNothing);
+
+        navigator.currentState.pushNamed(topRoute);
+        await tester.pump();
+
+        // Jump to halfway point of transition.
+        await tester.pump(const Duration(milliseconds: 150));
+        expect(find.byType(SharedAxisTransition), findsNWidgets(2));
+
+        // Rebuild the app without the transition.
+        await tester.pumpWidget(
+          MaterialApp(
+            navigatorKey: navigator,
+            home: const Material(
+              child: Text('abc'),
+            ),
+          ),
+        );
+        await tester.pump();
+        // Transitions should have been disposed of.
+        expect(find.byType(SharedAxisTransition), findsNothing);
+      },
+    );
+
+    testWidgets(
+      'State is not lost when transitioning',
+      (WidgetTester tester) async {
+        final GlobalKey<NavigatorState> navigator = GlobalKey<NavigatorState>();
+        const String bottomRoute = '/';
+        const String topRoute = '/a';
+
+        await tester.pumpWidget(
+          _TestWidget(
+            navigatorKey: navigator,
+            transitionType: SharedAxisTransitionType.scaled,
+            contentBuilder: (RouteSettings settings) {
+              return _StatefulTestWidget(
+                key: ValueKey<String>(settings.name),
+                name: settings.name,
+              );
+            },
+          ),
+        );
+
+        final _StatefulTestWidgetState bottomState = tester.state(
+          find.byKey(const ValueKey<String>(bottomRoute)),
+        );
+        expect(bottomState.widget.name, bottomRoute);
+
+        navigator.currentState.pushNamed(topRoute);
+        await tester.pump();
+        await tester.pump();
+
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(bottomRoute))),
+          bottomState,
+        );
+        final _StatefulTestWidgetState topState = tester.state(
+          find.byKey(const ValueKey<String>(topRoute)),
+        );
+        expect(topState.widget.name, topRoute);
+
+        await tester.pump(const Duration(milliseconds: 150));
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(bottomRoute))),
+          bottomState,
+        );
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(topRoute))),
+          topState,
+        );
+
+        await tester.pumpAndSettle();
+        expect(
+          tester.state(find.byKey(
+            const ValueKey<String>(bottomRoute),
+            skipOffstage: false,
+          )),
+          bottomState,
+        );
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(topRoute))),
+          topState,
+        );
+
+        navigator.currentState.pop();
+        await tester.pump();
+
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(bottomRoute))),
+          bottomState,
+        );
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(topRoute))),
+          topState,
+        );
+
+        await tester.pump(const Duration(milliseconds: 150));
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(bottomRoute))),
+          bottomState,
+        );
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(topRoute))),
+          topState,
+        );
+
+        await tester.pumpAndSettle();
+        expect(
+          tester.state(find.byKey(const ValueKey<String>(bottomRoute))),
+          bottomState,
+        );
+        expect(find.byKey(const ValueKey<String>(topRoute)), findsNothing);
+      },
+    );
+  });
 }
 
 double _getOpacity(String key, WidgetTester tester) {
@@ -1010,8 +1402,24 @@ double _getTranslationOffset(
         return a + translation.y;
       });
       break;
+    case SharedAxisTransitionType.scaled:
+      // SharedAxisTransitionType.scaled should not return a translation
+      // offset.
+      return null;
+      break;
   }
   return null; // unreachable
+}
+
+double _getScale(String key, WidgetTester tester) {
+  final Finder finder = find.ancestor(
+    of: find.byKey(ValueKey<String>(key)),
+    matching: find.byType(ScaleTransition),
+  );
+  return tester.widgetList(finder).fold<double>(1.0, (double a, Widget widget) {
+    final ScaleTransition transition = widget;
+    return a * transition.scale.value;
+  });
 }
 
 class _TestWidget extends StatelessWidget {
